+++ packages/core/src/posthog-core-stateless.ts
@@ -290,6 +290,9 @@ export abstract class PostHogCoreStateless {
     }
   }
 
+  /**
+   * @internal
+   */
   public addPendingPromise<T>(promise: Promise<T>): Promise<T> {
     return this.promiseQueue.add(promise)
   }