+++ packages/node/CHANGELOG.md
@@ -1,5 +1,22 @@
 # posthog-node
 
+## 5.26.0
+
+### Minor Changes
+
+- [#3144](https://github.com/PostHog/posthog-js/pull/3144) [`d43e90d`](https://github.com/PostHog/posthog-js/commit/d43e90df464b697a3aa0b46084d97a773a699abd) Thanks [@dustinbyrne](https://github.com/dustinbyrne)! - feat: getFeatureFlagResult, getAllFlags, getAllFlagsAndPayloads now have context-sensitive overrides which do not require `distinctId` as a parameter, instead reading it from the current context.
+  (2026-02-25)
+
+- [#3144](https://github.com/PostHog/posthog-js/pull/3144) [`d43e90d`](https://github.com/PostHog/posthog-js/commit/d43e90df464b697a3aa0b46084d97a773a699abd) Thanks [@dustinbyrne](https://github.com/dustinbyrne)! - feat: Add `enterContext()` to attach a context to the current async execution context
+  (2026-02-25)
+
+## 5.25.0
+
+### Minor Changes
+
+- [#3132](https://github.com/PostHog/posthog-js/pull/3132) [`301e75a`](https://github.com/PostHog/posthog-js/commit/301e75a8ecd7a983fc634d9ad9b521c8007a51d1) Thanks [@benjackwhite](https://github.com/benjackwhite)! - Officially adds support for super properties added with `.register` (it was available as a function but did not actually work previously)
+  (2026-02-23)
+
 ## 5.24.17
 
 ### Patch Changes